新規作成画面と編集画面のコンポーネントを統合
同じコードが2回書かれている箇所が多かったため、新規作成画面と編集画面のコンポーネントを統合しました。
