+++ src/components/MemoForm.vue
@@ -1,12 +1,28 @@
 <script setup>
 import MemoList from "./MemoList.vue";
 import { store } from "../store.js";
-import { useRouter } from "vue-router";
-const router = useRouter();
+import { computed } from "vue";
+import { useRouter, useRoute } from "vue-router";
 
-function onSubmit(event) {
+const route = useRoute();
+const router = useRouter();
+const editingMemo = computed(() => store.findMemo(route.params.id));
+
+function submitEditedMemo(event) {
   event.preventDefault();
-  store.add();
+  if (route.params.id) {
+    store.updateMemo(route.params.id);
+  } else {
+    store.add();
+  }
+  router.push("/");
+}
+
+function submitDeletedMemo(event) {
+  event.preventDefault();
+  if (route.params.id) {
+    store.deleteMemo(route.params.id);
+  }
   router.push("/");
 }
 </script>
@@ -15,9 +31,18 @@ function onSubmit(event) {
   <div class="memo-container">
     <MemoList />
     <form class="memo-form">
-      <textarea v-model="store.body" cols="30" rows="10"></textarea>
+      <textarea
+        v-if="editingMemo"
+        v-model="editingMemo.body"
+        cols="30"
+        rows="10"
+      ></textarea>
+      <textarea v-else v-model="store.body" cols="30" rows="10"> </textarea>
       <div class="memoapp-button">
-        <button type="submit" @click="onSubmit">保存</button>
+        <button type="submit" @click="submitEditedMemo">編集</button>
+        <button class="delete-button" type="submit" @click="submitDeletedMemo">
+          削除
+        </button>
       </div>
     </form>
   </div>
@@ -27,7 +52,9 @@ function onSubmit(event) {
 .memo-container {
   display: flex;
   align-items: flex-start;
-  justify-content: center;
+  width: 66%;
+  justify-content: space-between;
+  margin: 0 auto;
 }
 
 .memo-form {
