--- src/views/NewMemoView.vue
@@ -1,7 +1,0 @@
-<script setup>
-import NewMemoForm from "../components/MemoForm.vue";
-</script>
-
-<template>
-  <NewMemoForm />
-</template>
